+++ e2e/screenObject/BasicPaymentScreen.ts
@@ -85,25 +85,28 @@ class BasicPaymentScreen {
 
           button = $(`//input[@name='confirmAccountNumber']`);
           button.click();
-          button.sendKeys(['000123456789\n']);
-          driver.pause(2000);
-
-          button = $(`//input[@name='routingNumber']`);
-          button.click();
-          button.click();
-          button.sendKeys(['110000000\n']);
+          button.sendKeys(['000123456789']);
           driver.pause(2000);
 
           button = $(`//input[@name='accountNumber']`);
           button.click();
           button.click();
-          button.sendKeys(['000123456789\n']);
+          button.pressKeyCode;
+          button.sendKeys(['000123456789']);
           driver.pause(2000);
+
+          button = $(`//input[@name='routingNumber']`);
+          button.click();
+          button.click();
+          button.sendKeys(['110000000']);
+          driver.pause(2000);
+
+          button = $(`//input[@name='accountNumber']`);
+          button.click();
 
           button = $(`button*=Continue`);
           button.click();
-          button.click();
-          driver.pause(5000);
+          driver.pause(2000);
 
           button = $(`button*=Done`);
           button.click();